Changed 1 or 2 small things
If a researcher is not a user they won't have a profile so no information could be changed. Also each user should be able to see their own profile with their units and papers but other should not be able to see the units and papers that other users have earned and done respectively.